+++ src/main/java/courseitda/member/application/MemberService.java
@@ -1,12 +1,12 @@
 package courseitda.member.application;
 
-import org.springframework.stereotype.Service;
-
+import courseitda.auth.domain.AuthRole;
 import courseitda.member.domain.Member;
 import courseitda.member.domain.MemberRepository;
 import courseitda.member.ui.dto.request.SignUpRequest;
 import courseitda.member.ui.dto.response.SignUpResponse;
 import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
@@ -19,6 +19,7 @@ public class MemberService {
                 .nickname(request.nickname())
                 .email(request.email())
                 .password(request.password())
+                .authRole(AuthRole.MEMBER)
                 .build();
         final Member createdMember = memberRepository.save(member);
