+++ src/test/resources/application.yml
@@ -1,0 +1,30 @@
+spring:
+  application:
+    name: courseitda
+
+  datasource:
+    driver-class-name: org.h2.Driver
+    # DB_CLOSE_DELAY=-1: DB 연결이 끊어져도 메모리 유지
+    # DB_CLOSE_ON_EXIT=FALSE: JVM 종료 시까지 DB 유지
+    url: jdbc:h2:mem:testdb;MODE=MySQL;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
+    username: sa
+    password:
+
+  jpa:
+    hibernate:
+      ddl-auto: create-drop
+    properties:
+      hibernate:
+        format_sql: true
+        dialect: org.hibernate.dialect.H2Dialect
+    show-sql: true
+
+  h2:
+    console:
+      enabled: false
+
+security:
+  jwt:
+    access-token:
+      secret-key: since-jjwt-api-0.11-secret-key-must-be-longer-than-32-bytes-for-test
+      validity-in-milliseconds: 3600000 # 1 hour
